+++ src/main/java/io/odpf/firehose/sink/mongodb/util/MongoSinkFactoryUtil.java
@@ -1,6 +1,10 @@
 package io.odpf.firehose.sink.mongodb.util;
 
+import com.mongodb.MongoClient;
+import com.mongodb.MongoClientOptions;
+import com.mongodb.MongoCredential;
 import com.mongodb.ServerAddress;
+import io.odpf.firehose.config.MongoSinkConfig;
 import io.odpf.firehose.metrics.Instrumentation;
 import lombok.experimental.UtilityClass;
 
@@ -9,6 +13,9 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+/**
+ * The type Mongo sink factory util.
+ */
 @UtilityClass
 public class MongoSinkFactoryUtil {
 
@@ -36,5 +43,40 @@ public class MongoSinkFactoryUtil {
             instrumentation.logError("No connection URL found");
             throw new IllegalArgumentException("SINK_MONGO_CONNECTION_URLS is empty or null");
         }
+    }
+
+    /**
+     * Gets status codes as list.
+     *
+     * @param mongoRetryStatusCodeBlacklist the mongo retry status code blacklist
+     * @return the status codes as list
+     */
+    public static List<String> getStatusCodesAsList(String mongoRetryStatusCodeBlacklist) {
+        return Arrays
+                .stream(mongoRetryStatusCodeBlacklist.split(","))
+                .map(String::trim)
+                .filter(s -> !s.isEmpty())
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * Builds the Mongo client.
+     *
+     * @param mongoSinkConfig the mongo sink config
+     * @param instrumentation the instrumentation
+     * @return the mongo client
+     */
+    public static MongoClient buildMongoClient(MongoSinkConfig mongoSinkConfig, Instrumentation instrumentation) {
+        List<ServerAddress> serverAddresses = MongoSinkFactoryUtil.getServerAddresses(mongoSinkConfig.getSinkMongoConnectionUrls(), instrumentation);
+        MongoClientOptions options = MongoClientOptions.builder().connectTimeout(mongoSinkConfig.getSinkMongoRequestTimeoutMs()).build();
+
+        MongoClient mongoClient;
+        if (mongoSinkConfig.isSinkMongoAuthEnable()) {
+            MongoCredential mongoCredential = MongoCredential.createCredential(mongoSinkConfig.getSinkMongoAuthUsername(), mongoSinkConfig.getSinkMongoAuthDB(), mongoSinkConfig.getSinkMongoAuthPassword().toCharArray());
+            mongoClient = new MongoClient(serverAddresses, mongoCredential, options);
+        } else {
+            mongoClient = new MongoClient(serverAddresses, options);
+        }
+        return mongoClient;
     }
 }
